refactor: 분리된 DTO와 interface들을 사용하도록 경로 수정
